requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
SQLAlchemy==1.4.32
Flask-SQLAlchemy==2.5.1
Flask-Bcrypt==0.7.1
Flask-JWT-Extended==4.4.2
pycryptodome==3.21.0